Deepfake Detection/requirements.txt
Django==5.0.2
numpy==1.26.4
opencv_contrib_python==4.9.0.80
Pillow==10.3.0
transformers==4.38.2
